+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Modified the screen size of the example #3
 - Changed the max sound length to 48
 - Changed the max music length to 32
+- Added the playing property
+- Added the play_pos function
+
 
 ## 1.0.1
 - Simplified comparisons